+++ src/models/index.ts
@@ -26,6 +26,8 @@ export interface Meta {
   failed: boolean;
   /** The date/time in ISO-8601 format when the document will be automatically deleted.  Defaults to no expiry. */
   expiryTime?: string;
+  /** The resume's language. */
+  language?: string;
 }
 
 export interface RequestError {
@@ -49,6 +51,8 @@ export interface ResumeData {
   objective?: string;
   /** NOTE: This property will not be serialized. It can only be populated by the server. */
   readonly languages?: string[];
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly languageCodes?: string[];
   summary?: string;
   totalYearsExperience?: number;
   /**
@@ -150,7 +154,13 @@ export interface ResumeDataEducationItemDates {
 export interface ResumeDataWorkExperienceItem {
   id?: number;
   jobTitle?: string;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly socCode?: string;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly socName?: string;
   organization?: string;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly industry?: string;
   location?: Location;
   jobDescription?: string;
   dates?: ResumeDataWorkExperienceItemDates;
@@ -189,18 +199,30 @@ export interface Components1TryetgSchemasResumedataPropertiesWorkexperienceItems
 
 export interface ResumeDataSkillsItem {
   id?: number;
+  /**
+   * EMSI id of this skill.
+   * NOTE: This property will not be serialized. It can only be populated by the server.
+   */
+  readonly emsiId?: string;
   name?: string;
   lastUsed?: string;
   numberOfMonths?: number;
   /** NOTE: This property will not be serialized. It can only be populated by the server. */
   readonly type?: string;
   /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly count?: number;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
+  readonly weighting?: number;
+  /** NOTE: This property will not be serialized. It can only be populated by the server. */
   readonly sources?: ResumeDataSkillsPropertiesItemsItem[];
 }
 
 export interface ResumeDataSkillsPropertiesItemsItem {
   section?: string;
+  /** If this skill is extracted from a "workExperience" section, the "position" is the index of the work experience where this skill is found, with 0 being the first work experience, 1 being the second work experience, and so on. */
   position?: number;
+  /** If this skill is extracted from a "workExperience" section, the "workExperienceId" is the id of the work experience where this skill is found. */
+  workExperienceId?: number;
 }
 
 export interface ResumeDataRefereesItem {
@@ -1459,6 +1481,8 @@ export interface AffindaAPIGetResumeSearchMatchOptionalParams
   extends coreClient.OperationOptions {
   /** Optionally, specify an index to search in. If not specified, will search in all indexes. */
   index?: string;
+  /** Add keywords to the search criteria. */
+  searchExpression?: string;
   /** How important is this criteria to the matching score, range from 0 to 1. */
   jobTitlesWeight?: number;
   /** How important is this criteria to the matching score, range from 0 to 1. */